requirements.txt
```

Paste this:
```
flask==3.0.0
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
schedule==1.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
